Update readgrid & change grid size
The readgrid subroutine has been updated. It now reads a file called "grid.txt" and will initialize values of the grid accordingly. Also, the space allocated for grid has been updated to 4000 bytes instead of 400.

I realize I have put this change in a branch named printgrid but did not know how to change the name.